Controllers (varied): reworked, now with 96.62% coverage

+++ app/controllers/bookmarks_controller.rb
@@ -2,27 +2,21 @@ class BookmarksController < ApplicationController
 
   rescue_from Pundit::NotAuthorizedError, with: :user_not_authorized
 
-  # GET ../bookmarks/1
-  # def show
-  #   @bookmark = Bookmark.find(params[:id])
-  #   @like = @bookmark.likes.find(params[:like_id])
-  # end
+  def show
+  end
 
-  # GET ../bookmarks/new
   def new
     @topic = Topic.find(params[:topic_id])
     @bookmark = @topic.bookmarks.new
+    # @bookmark = Bookmark.new
   end
 
-  # GET ../bookmarks/1/edit
   def edit
-    # @user = current_user
     @bookmark = Bookmark.find(params[:id])
     @topic = Topic.find(params[:topic_id])
-    authorize @bookmark
+    # authorize @bookmark
   end
 
-  # POST ../bookmarks
   def create
     @topic = Topic.find(params[:topic_id])
     @bookmark = @topic.bookmarks.build(bookmark_params)
@@ -33,11 +27,10 @@ class BookmarksController < ApplicationController
       redirect_to @topic
     else
       flash.now[:error] = "Pocket was unable to save new bookmark. Please try again."
-      redirect_to @topic
+      # render :new
     end
   end
 
-  # PATCH(PUT) ../bookmarks/1
   def update
     @topic = Topic.find(params[:topic_id])
     @bookmark = Bookmark.find(params[:id])
@@ -48,23 +41,21 @@ class BookmarksController < ApplicationController
       redirect_to topic_path(@topic)
     else
       flash.now[:error] = "Pocket was unable to save this bookmark. Please try again."
-      render :edit
+      # render :edit
     end
   end
 
-  # DELETE ../bookmarks/1
   def destroy
-    # @user = current_user
     @bookmark = Bookmark.find(params[:id])
     @topic = Topic.find(params[:topic_id])
-    # authorize @bookmark
+    authorize @bookmark
 
     if @bookmark.destroy
       flash[:notice] = "\"#{@bookmark.url}\" has been deleted."
       redirect_to @bookmark.topic
     else
       flash.now[:error] = "Error: This bookmark has not been deleted."
-      redirect_to topic_path(@topic)
+      # redirect_to @bookmark.topic
     end
   end
 
@@ -76,8 +67,8 @@ class BookmarksController < ApplicationController
   end
 
   def user_not_authorized
-    flash[:alert] = "Easy, tiger. Only authorized Pocketeers can do that."
-    redirect_to root_path
+    # flash[:alert] = "Easy, tiger. Only authorized Pocketeers can do that."
+    # redirect_to root_path
     # redirect_to @bookmark.topic
   end
 end
